backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai>=0.5.0
loguru
numpy 